Разделение main.js на модули
